Use Adwaita widgets in status editor
The `PidginStatusChooser` is also changed from a `GtkDropDown` to a `AdwComboRow` as everything that uses it is now in an `AdwPreferencesGroup`.

Also simplifies a bit of the status editor by using responses instead of widgets.

Testing Done:
Compiled and opened Statuses, modified an existing one and confirmed the dialog was reloaded correctly and appeared to work.

Reviewed at https://reviews.imfreedom.org/r/2396/