Don't duplicate contacts in the demo protocol when reconnecting
Previously, we were always creating new contacts and persons on connection. This
ends up with duplicates when we run the connection error actions which is
obviously not what we want.

Testing Done:
Ran the connection error actions and verified that we did not get duplicate contacts and persons.

Reviewed at https://reviews.imfreedom.org/r/2436/