Set _POSIX_C_SOURCE as we need it for sigemptyset when using muslc
muslc requires _POSIX_C_SOURCE to be set to get sigemptyset and a few other
things. I went with 200809L as the value since we're using the C17 standard now
and this is the most recent POSIX version that I could find.

Testing Done:
Ran the `alpine-edge-amd64` convey plan locally and verified the build was successful.

Reviewed at https://reviews.imfreedom.org/r/2480/