Add purple_tags_get_all_with_name
I realized I needed this when I started porting XMPP to the new Contact API. I
need to be able to get a list of all the groups that a contact is in so I can
updated them appropriately but we didn't have the api for it. This is something
that other protocols will need to do as well so here's the API.

Testing Done:
Ran the unit tests.

Reviewed at https://reviews.imfreedom.org/r/2503/