Require GTK 4.10 and GLib 2.76
Due to the deprecations in GTK 4.10 we're moving to it immediately so that we
don't end up having to deal with deprecation warnings forever (again).

Also, we're taking this opportunity to require the last glib minor version as
well as if you have GTK 4.10, you most likely have GLib 2.76 as well.

Testing Done:
Compiled.

Reviewed at https://reviews.imfreedom.org/r/2526/