Remove the spellchk plugin
While this is a good idea, the implementation leaves a bit to be desired in
2023. First of all, this should probably be integrated with spell checking,
which unfortunately is currently an open problem in GTK4. Also we don't
currently have a way for plugins to provide their own preferences UI.

Testing Done:
Asked the turtles to check it out, and they said everything was cool.

Reviewed at https://reviews.imfreedom.org/r/2541/