Fix a number of bugs in the account editor
This fixes the protocol not being the first one in the list as well as the
username still being populate from a previous account when a new account is
being created.

All login options are now set to insensitive when an account is online and are
properly made sensitive when the account is offline. However, these properties
aren't bound yet, because there's no property to bind to. See PIDGIN-17815.

Testing Done:
Opened the account editor for an existing online account and verified that the protocol, username, and all user splits were insensitive. Hit the back button, hit the add button and verified that the protocol was at Bonjour (the first in the list) and that the username was empty and that those 2 and all of the user splits were sensitive.

Bugs closed: PIDGIN-17813

Reviewed at https://reviews.imfreedom.org/r/2540/