Don't use the Real name as a candidate for the SASL username in IRC
The real name field is free form text which means it can contain characters
that aren't allowed in irc/sasl account names. So if a user has a real name set
but not a sasl login name, they get hard to trouble shoot authentication
failures. So to avoid all that, we're removing real name as an option and just
allowing the "newer" field and falling back to the username if that field isn't
set.

Testing Done:
Compiled

Reviewed at https://reviews.imfreedom.org/r/2535/