Remove the notify plugin
This hasn't been building by default for a while and depends on a whole bunch
of old conversation window api that it is the only consumer of. Also most people
nowadays are used to just getting a badge which is what we're planning on doing in
the near future. For all these reasons, we should just delete this.

Testing Done:
Ran `ninja turtles` without issue.

Reviewed at https://reviews.imfreedom.org/r/2537/