Remove a bunch of dead code from gtkconv
I tried my best to get the keybindings working, but none of them did.

Testing Done:
Compiled and tried all sorts of keybindings in the history and input. I had a message selected in the history and was able to type in the input without any of that old code.

Reviewed at https://reviews.imfreedom.org/r/2538/