Set the sasl login name on successful login if not set
We tried to do this previously, but it was just done when the connection was
created rather than waiting until we had successfully logged in with the user's
nick.

Testing Done:
Created an ircv3 account with the wrong user name, failed to login, and made sure that the sasl login name was not set. Then fixed the username, logged in successfully, and verified that the sasl login name was set properly.

Reviewed at https://reviews.imfreedom.org/r/2559/