Stop ignoring SIGCHLD in the unit tests
This was breaking the unit tests with glib 2.77 and isn't even necessary any
more now that we use GResolver.

This was originally thought to be an upstream [GLib bug](https://gitlab.gnome.org/GNOME/glib/-/issues/3071) which it kind of is, as they made changes, but this code shouldn't exist anyways.

Testing Done:
Ran the unit tests and verified they passed. Also connected an ircv3 account and verified it worked fine.

Reviewed at https://reviews.imfreedom.org/r/2565/