IRCv3: work around some warnings when generating the gir file
Testing Done:
I generated the gir with and without this and diffed the versions and there were no changes other than that the warnings were gone.

I suspect this is a bug somewhere in either `g-ir-scanner` or `GObject` but this gets rid of the warning for now which will stop others from wasting time on it for the time being.

Reviewed at https://reviews.imfreedom.org/r/2582/