Don't try to disconnect account when finalizing protocols
Accounts that are connected have a reference to a protocol, so this code will
never actually be called as the protocol is still referenced.

Testing Done:
Connected an xmpp account and exited without issue. I tried looking at valgrind output, but there's a ton of it that should probably be addressed separately.

Reviewed at https://reviews.imfreedom.org/r/2583/