Rename PidginConversation to PidginConversationOld
I'm working on a replacement for PidginConversation and want to use the name
while building it without actually changing anything. So to do that, we need to
rename the old one first.

Testing Done:
Connected and IRCv3 account and verified the status window worked.

Reviewed at https://reviews.imfreedom.org/r/2587/