Add a "messages" property to PurpleConversation
This is a GListModel that can be used in the ui's for handling history. Right
now it is a simple GListStore, but the idea is that in the future we will have
a PurpleHistoryListStore or something that will only keep a subset of the
messages in memory and then update according to the backing store when
necessary.

Testing Done:
Ran the unit tests.

Reviewed at https://reviews.imfreedom.org/r/2590/