A basic context menu for PurpleContactInfo
This is currently only displayed via the contact list and it takes a shortcut
to just show the menu for the priority contact.

The view profile option currently just logs the profile as we need to determine
how we want the user interface to look/work for this yet.

Testing Done:
Verified that `view profile` was disabled for an xmpp account.
Verified that `view profile` worked for a demo contact with a profile and one without.

Bugs closed: PIDGIN-17682

Reviewed at https://reviews.imfreedom.org/r/2592/