Add the conversation as a parameter to PurpleProtocolChat->send
This is an incremental update so that all of the implemenations can start
moving in this direction. Eventually this be something like
PurpleProtocol->send_message or something which will be used for both ims and
chats.

Testing Done:
Joined an XMPP muc and sent some messages.

Bugs closed: PIDGIN-17826

Reviewed at https://reviews.imfreedom.org/r/2596/