Add color and color-for-display properties to PurplePerson
The color property allows setting a custom color for PurplePerson that should
be used whenever displaying the person or a contact from that person.

To make this easier the color-for-display property should be used as it will
fallback to the priority contact if the person does not have a color set.

Testing Done:
Ran the unit tests under valgrind.

Bugs closed: PIDGIN-17817

Reviewed at https://reviews.imfreedom.org/r/2602/