Turn most of the ChatConversation API into wrappers for the PurpleConversation methods
This doesn't handle membership yet as that's going to be kind of big.

Also removed all of the ignoring stuff for the moment as that is not currently
implemented anywhere and needs to fallback to the conversation class at
some point.

Testing Done:
Joined an XMPP MUC and sent some messages.
Also verified that topic changing worked correctly.

Reviewed at https://reviews.imfreedom.org/r/2597/