Create PurpleChannelJoinDetails to replace the components for join chats
This will be used be a new PurpleProtocolConversation vfunc in the near future
which the ui can request to get information about how to join a chat for a
protocol.

Testing Done:
Ran the unit tests under valgrind and verified the docs built with no new warnings.

Reviewed at https://reviews.imfreedom.org/r/2616/