Add tooltips to timestamps in PidginConversation
This allows you to get a tooltip for the timestamp and the edited label that
displays the time in the locale's default format. The edited label was tested
by forcing the demo protocol plugin to mark messages as edited.

Testing Done:
Sent a bunch of messages with the echo bot and verified the timestamp tooltip on IRCv3 as well.

Reviewed at https://reviews.imfreedom.org/r/2645/