Remove the buddy notes plugin
Storage of this has been moved to PurpleContactInfo as some protocols support
this natively. Removing this now means we lose the ui portion for this, but
that's okay for now as this should be integrated into the default menu for
ContactInfo.

Testing Done:
Asked the turtles for help.

Reviewed at https://reviews.imfreedom.org/r/2647/