Fix postconf script when used as a subproject
We add the Pidgin root to the include path, but saved the `meson-config.h` header to the global build root. This is normally fine when building Pidgin standalone, but when Pidgin is a subproject, those are not the same place, and the header cannot be found.

Additionally, attempt to strip out all options that are not directly for Pidgin. For a full-subproject build, the option string is 157KB, which is too large to use as a C string, so we need to cut back on this a bit.

Testing Done:
Compiled with Pidgin as a subproject and not, then checked the generated `meson-config.h` header seemed reasonable.

Reviewed at https://reviews.imfreedom.org/r/2653/