Rework the info pane a bit
This removes the presence icon as the plan is to work that into a reworked
PidginAvatar. The PidginAvatar is now the first item in the info pane. Also,
hide the topic label if no topic is set, and center align the conversation
name and topic.

Testing Done:
Opened some dm's and testing@conference.pidgin.im and verified everything looked okay.

Reviewed at https://reviews.imfreedom.org/r/2649/