Add a get_history_adapter method to PurpleUi and implement in Pidgin and Finch
This should simplify the process of starting up the history API as it puts the
work into libpurple which will fallback to an in-memory history adapter if the
UI did not provide one.

Testing Done:
Ran the unit tests under valgrind.
Also sent some messages to the echo bot with Pidgin.

Reviewed at https://reviews.imfreedom.org/r/2738/