Port TalkatuAutoScroller to PidginAutoAdjustment
The name change makes more sense so I went with it. I also moved the
documentation to the header files to keep with Pidgin's style.

Testing Done:
Consulted with my friends the turtles.
Sent a bunch of messages to the echo bot to test the auto scrolling abilities.
Verified the class name in the GTK Inspector.

Reviewed at https://reviews.imfreedom.org/r/2742/