Create pidgin_text_buffer API for working with GtkTextBuffers
Currently we only implement _get_html and _get_html_range and only a subset of
html tags are identified properly. This was intentional as it's easy to add
additional tags and because we're also introducing testing to Pidgin in this
commit.

Testing Done:
Ran the unit tests under valgrind without issues.

Reviewed at https://reviews.imfreedom.org/r/2754/