Integrate PurplePresenceManager with the core and uis
The original implemenation was just the API and wasn't yet tied into the core
or ui. This fixes that.

Testing Done:
Temporarily added code to add a saved status in the `PurpleUi->get_presence_manager` implementation and verified it was saved to the correct file.

Reviewed at https://reviews.imfreedom.org/r/2771/