Simplify PidginInfoPane by using bindings everywhere
I also had to update purple_markup_linkify to use lower case tags for pango.
That said, I'm not sure why they were upper case in the first place, that might
be _very_ old.

I played around with a tooltip on the topic since it can be ellipsized but you
can't click the links in a tooltip so that was a bust. I also tried making the
top label expand on hover but that didn't work well either. So we'll most
likely end up putting the full topic in a "Conversation Info" pane that'll swap
out with the member list and stuff.

Testing Done:
Checked in with the turtles. Also opened a dm and a channel without issue.

Reviewed at https://reviews.imfreedom.org/r/2845/