Add purple_contact_find_dm
This is a common use case and is a bit of code, so it's good to have a helper
like this.

Testing Done:
Created dm's via the contact list and made sure they worked. The previous update /r/2851 broke instant messages for all protocol plugins that are still using PurpleIMConversation which need to be updated anyways.

Reviewed at https://reviews.imfreedom.org/r/2853/