Updated purple_conversation_remove_member to take a PurpleContactInfo
All of the purple_conversation_*_member methods took a PurpleContactInfo except
for remove which this addresses. This makes the API more consistent and easier
to use as the protocols will generally have a PurpleContactInfo but not a
PurpleConversationMember when doing member management. That additional lookup
caused us to have to iterate the list twice where now it's only one.

Testing Done:
Ran the unit tests

Reviewed at https://reviews.imfreedom.org/r/2844/