Remove gtkblist
The only thing that was still being used here was the join chat dialog which
we're migrating everything over to join channel so it didn't make sense to
keep it around.

Testing Done:
Consulted with the turtles and connected a demo and ircv3 account.

Reviewed at https://reviews.imfreedom.org/r/2893/