Remove gtkconv.[ch]
The only thing still using this is the unity plugin which will need to be
updated to the new api when it is finalized. In the meantime, I just disabled
it by default so we can remove all of this.

Testing Done:
Ran with the turtles and did a quick connectivity check.

Reviewed at https://reviews.imfreedom.org/r/2895/