Remove unused actions from the display window
We removed a bunch of these previously as they were context dependent and we
want to make those more obvious in the future.

For the actions that were the only user of old dialog functions, this commit
deletes those functions as well.

Testing Done:
Consulted with turtles and verified a few accounts connected without issue.

Reviewed at https://reviews.imfreedom.org/r/2897/