Remove some left over code for enabling and disabling actions that is no longer necessary
When we removed the unused menu items we missed removing the logic for toggling
whether or not actions that required the PurpleProtocolChat interface. This
just deletes that code.

Testing Done:
Compiled, ran, and verified that the warning about the unknown action was gone.

Reviewed at https://reviews.imfreedom.org/r/2904/