Implement PurpleFileTransferManager
This is a pretty simple manager as duplicates are allowed and
PurpleFileTransfer all property driven which means we only need to propagate
the notify signal.

This doesn't persist across restarts as we have some other issues that are
making that harder than it needs to be.

Testing Done:
Ran the unit tests under valgrind.

Bugs closed: PIDGIN-17831

Reviewed at https://reviews.imfreedom.org/r/2874/