Create a full widget for the new instant message window
This now uses AdwMessageDialog and is self contained but PidginApplication
manages a singleton of it.

Previously this used the request api. I'm not sure why, as the request fields
were created by the ui itself.

Testing Done:
Created an im with `Echo` on a `demo` account and was able to send and receive messages.

Reviewed at https://reviews.imfreedom.org/r/2896/