Add the CONSTRUCT flag to PurpleContactInfo:id as we need it set in the constructor
I found this while trying to rewriting the account serialization code. The id
was being set multiple times when it was passed to g_object_new as that value
was not set before GObjectClass.constructed was called.

Testing Done:
Checked in with turtles, but also used work in progress branch to verify the fix.

Reviewed at https://reviews.imfreedom.org/r/3008/