Explicitly cast libxml error handler function
Depending on compiler options, the `const` → non-`const` argument can fail in just the same way as it did in reverse. So add an explicit cast to silence that.

Testing Done:
Compiled with `gcc` and `clang` on a system with older libxml2, which still had the non-`const` type. There, `gcc` raises a warning, but `clang` seems to think this is a fatal error.

Reviewed at https://reviews.imfreedom.org/r/3016/