Fix transfer annotation of purple_notification_manager_add
This function either returns immediately (for a double add), or adds a ref
before putting it in the `GPtrArray`. So it is `transfer none`, not `transfer full`.

Testing Done:
Ran `ninja turtles` and also tested with valgrind.
Also added notifications from the demo protocol in valgrind.

Reviewed at https://reviews.imfreedom.org/r/3022/