Stop saving notification setting on unload
When libpurple is shutting down, the settings backend is gone before the plugins (and maybe this should be fixed elsewhere), so all the saved settings leak as they attempt to write to a `NULL` settings object.

All places in this plugin that change the settings then save them, so there's no need to save on unload.

Testing Done:
Ran in valgrind with the plugin enabled, and there were no longer any leaks of settings values from unload at exit.

Reviewed at https://reviews.imfreedom.org/r/3024/