Replace purple_accounts_restore_statuses with purple_account_manager_set_online
This is a step in replacing the old accounts API. I skipped checking if we have
network connectivity or not, because we need to work that into
purple_account_connect and checking purple_protocol_can_connect_async.

Testing Done:
Ran with the turtles.

Reviewed at https://reviews.imfreedom.org/r/3002/