Add missing documentation for all of the media API
I had to guess on some of this, and I even noted it sometimes, but this makes
the doc check happy for now until we come back around and refactor this API.

Testing Done:
Ran the unit tests with the doc check. That's not include in this as this is already a lot of changes.

Reviewed at https://reviews.imfreedom.org/r/3044/