Remove the GaduGadu protocol plugin
This protocol plugin depends on libgadu which hasn't been maintained since
2019. That said it is still currently functional as far as I know. However, the
protocol plugin needs a lot of work for the changes in purple 3 and at this
point, it is probably easier to rewrite it even if it still uses libgadu.

Testing Done:
Ran meson dist.

Reviewed at https://reviews.imfreedom.org/r/3051/