Remove the old Bonjour protocol plugin
We have a started a new version and with migrations that are currently in place
this didn't really work anymore anyways. So rather than letting it sit around
and make things harder to migrate we're removing it.

Testing Done:
Hung out with the turtles.

Reviewed at https://reviews.imfreedom.org/r/3072/