Remove Finch
We are intending on moving Finch to its own repository, but that's going to
take some time and we have other things to do at the moment. So instead we're
removing it so that we can move forward with everything else immediately.

Testing Done:
Hung out with the turtles

Reviewed at https://reviews.imfreedom.org/r/3076/