Update PurpleAccountManager to connect/disconnect accounts when enabled/disabled
I'm not sure why we didn't do this before, but now that we have the online
property it makes much more sense for the account manager to actually connect
and disconnect the accounts.

Testing Done:
Bought the turtles pizza and toggled a demo account a few times.

Reviewed at https://reviews.imfreedom.org/r/3098/