Remove PurpleChatConversation usage from Pidgin
PurpleChatConversation is on its way out, but this is blocking it's removal.

The invite dialog was also removed as it will need to be updated to be aware
of adding users to existing conversations that could be group dms or even
channels.

Testing Done:
Had the turtles run this on top of the branch from r/3106

Reviewed at https://reviews.imfreedom.org/r/3105/