Update the demo protocol plugint to not use serv_got_im
serv_got_im is going away so this needs to stop using it. For the interim, I
just made is use g_warning instead.

Testing Done:
Ran the turtles and built on top of a branch that's removing a lot of the old im stuff.

Reviewed at https://reviews.imfreedom.org/r/3107/