Stop using PurpleIMConversation in Pidgin
We're trying to remove PurpleIMConversation so we need to stop using it in
Pidgin. There weren't many uses that were still being called so this was pretty
straight forward.

Testing Done:
Ran with the turtles and dinked around in the app a bit.

Reviewed at https://reviews.imfreedom.org/r/3112/