Remove request usage from PurpleAccount
The request API is on its way out and this needed to be cleaned up. There is
still a use case for changing passwords, but that should probably be added to
PurpleUi. The profile updating will most likely be moved to PurpleProtocol but
we need to determine that in the future.

Testing Done:
Checked in with the turtles and connected a demo account.

Reviewed at https://reviews.imfreedom.org/r/3146/