Prepare for reparenting PurpleAccount
The idea of having PurpleAccount subclass PurpleContactInfo had some issues.
First, the server side id isn't known at account creation time which means we
had to generate a local one and then throw that away when we knew the server
side id. This technically worked, but breaks the current idea for serialization
which is one of the big reasons why this has to change.

This is the first step in reparenting it to GObject and having a property that
has the protocol specific contact info. The idea is that we can update all of
the consumers to call these methods and once that's done we can go ahead and do
the reparenting without breaking anything.

Testing Done:
Had the turtles do their thing and connected some accounts without issue.

Reviewed at https://reviews.imfreedom.org/r/3157/