Automatically generate titles for DMs and GroupDMs
This is one of the steps in removing PurpleConversation:name. This required a
but of tinkering with property flags and freezing the property notification,
but I think it's all good.

Testing Done:
Ran the turtles, ran the tests under valgrind, and opened conversations with the demo protocol plugin and sent some messages.

Reviewed at https://reviews.imfreedom.org/r/3158/