Add all of the demo prpl's contacts to contacts.json
Now that creating conversations depends on PurpleContactInfo's we needed to add
contact info's for all of the users that are used in the "trigger a contact
adding you" action to make the message button on the notification work.

Testing Done:
Called in the turtles and made sure the message button worked for everyone.

Reviewed at https://reviews.imfreedom.org/r/3163/