Some cleanups for PidginDisplayItem
Set the title labels to ellipsize at the end, fix the position of the
notification indicators, and hide the notification indicators when the item is
selected. We're hiding them because a later review request is going to allow
us to show additional widgets on hover and selection.

Testing Done:
Ran the turtles and opened a bunch of conversations and notifications. See the screen shots in the review.

Reviewed at https://reviews.imfreedom.org/r/3177/