IRCv3: request the account-tag capability if it's available
Right now we just request this because ideally we will use this value as the `id` in `PurpleContactInfo` but that means nicks need to be a display name, and that is going to take a bit more work to work through.

Testing Done:
Connected to my local ergo and verified that my account and other registered accounts showed up in the `account` tag for their messages.

Reviewed at https://reviews.imfreedom.org/r/3252/