Remove purple_normalize
This shouldn't really be necessary in the future as we're working with contacts
and protocols can adjust as necessary. Also the last use was in the account
manager to help determine if a duplicate account existed by username, but
usernames will be going away/changing in the future so this would just be an
issue then too.

Testing Done:
Ran the turtles.

Reviewed at https://reviews.imfreedom.org/r/3367/