Remove Purple.ProtocolActions
Everything has been migrated to Purple.Protocol.get_action_menu except for the
actions-changed signal which we'll revisit when we are ready for it as I'm not
sure the current or previous setup would have actually worked.

Testing Done:
Called in the turtles and verified the actions still work on the demo protocol plugin.

Bugs closed: PIDGIN-17884

Reviewed at https://reviews.imfreedom.org/r/3396/