Use Purple.Account.username in cases where Purple.ContactInfo.name_for_display is NULL
This solves the following situations where an account name was displayed as "(null)":
* in the confirmation dialog, when removing an account from the account manager
* in the account combo box in the join channel dialog

Testing Done:
Removed an account from the account manager and verified that the account name was not "(null)" in the confirmation dialog.
Opened the Join Channel dialog with a connected IRC account and verfied that the account name was not "(null)".

Reviewed at https://reviews.imfreedom.org/r/3431/