Add a logging property to Purple.Conversation
This property is used to determine whether or not we are logging messages.
It's setup in a way to mimic the way log toggling works in Google Chat so that
can be replicated, but our users may want to ignore that in the future.

Testing Done:
Ran the conversation tests under valgrind and let the turtles handle the rest.

Reviewed at https://reviews.imfreedom.org/r/3441/