Add Purple.Conversation:drafting
This allows user interfaces to track whether a user has started drafting a
message in a conversation and can then display something to show that.

Testing Done:
Ran the tests under valgrind and called in the turtles for the rest.

Bugs closed: PIDGIN-17916

Reviewed at https://reviews.imfreedom.org/r/3452/