Display CTCP requests in the status window
In the future we might show this in conversations if they exist, but for now
this gets the job done. The content of the message is the same thing we
displayed in Pidgin 2.

Testing Done:
Connected to my local ergo and sent CTCP messages from pidgin 2 and irssi to my pidgin 3 instance and a channel.

Also called in the turtles for good measure.

Bugs closed: PIDGIN-17893

Reviewed at https://reviews.imfreedom.org/r/3470/