Create Pidgin.Notification to simplify the custom widgets
This should be much easier to maintain as it keeps everything encapsulated.

PidginNotificationAuthorizationRequest was removed as nothing creates these
notifications right now.

PidginNotificationConnectionError was removed as we are currently in the
process of reworking the way connection errors work and this is in the way of
that.

Testing Done:
Created a bunch of test notifications with the demo protocol plugin and called in the turtles.

Bugs closed: PIDGIN-17929, PIDGIN-17942

Reviewed at https://reviews.imfreedom.org/r/3480/