Shorten plugin load error notification title
Using the full path can create a really long title, which takes up a lot of
space unnecessarily. These details can go in the subtitle (and I assume we want
to keep the full path for the user to be able to investigate.)

Testing Done:
Opened notifications with an invalid plugin in the saved list, and the title was shorter.

Reviewed at https://reviews.imfreedom.org/r/3501/